server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.10.0
appdirs==1.4.4
asgiref==3.9.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.4
blinker==1.9.0
browserbase==1.4.0
build==1.3.0
cachetools==6.2.1
certifi==2025.8.3
cffi>=1.15.0
channels==4.3.1
charset-normalizer==3.4.3
chromadb==1.1.1
click==8.3.0
coloredlogs==15.0.1
crewai==0.203.1
crewai-tools==0.76.0
cryptography==46.0.2
curl_cffi==0.13.0
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docker==7.1.0
docstring_parser==0.17.0
dotenv==0.9.9
durationpy==0.10
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.20.0
filetype==1.2.0
finnhub-python==2.4.25
flatbuffers==25.9.23
frozendict==2.4.6
frozenlist==1.7.0
fsspec==2025.9.0
google-ai-generativelanguage==0.7.0
google-api-core==2.26.0
google-auth==2.41.1
google-genai==1.43.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.11.3
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
lance-namespace==0.0.18
lance-namespace-urllib3-client==0.0.18
lancedb==0.25.2
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.79
langchain-google-genai==2.1.12
langchain-huggingface==0.3.1
langchain-openai==0.3.32
langchain-pinecone==0.2.12
langchain-text-splitters==0.3.9
langsmith==0.4.13
litellm==1.74.9
load-dotenv==0.1.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.3
multitasking==0.0.12
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.2
oauthlib==3.3.1
onnxruntime==1.23.1
openai==1.109.1
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.1
overrides==7.7.0
packaging==24.2
pandas==2.3.1
parso==0.8.5
pdfminer.six==20250506
pdfplumber==0.11.7
peewee==3.18.2
pexpect==4.9.0
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.8
playwright==1.55.0
portalocker==2.7.0
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic>=2.11.9,<3
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pylance==0.38.2
pypdf==6.1.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytube==15.0.0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
referencing==0.37.0
regex==2025.9.18
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.1
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
sqlparse==0.5.3
stack-data==0.6.3
stagehand==0.5.4
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tomli_w==1.2.0
torch==2.8.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uv==0.9.2
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
yarl==1.20.1
yfinance==0.2.65
youtube-search==2.1.2
youtube-transcript-api==1.2.3
zipp==3.23.0
zstandard==0.23.0
